move table handling to separate file
